requirements.txt
# Dev dependencies
nose
rednose
pyyaml
